+++ test/components/slider/slider.e2e-spec.js
@@ -67,3 +67,34 @@ describe('Slider short tests', () => { //eslint-disable-line
     });
   }
 });
+
+describe('Slider tooltip position test', () => { //eslint-disable-line
+  beforeEach(async () => {
+    await utils.setPage('/components/slider/example-tooltip-position');
+  });
+
+  it('Should show the tooltip on top', async () => {
+    const sliderEl = await element(by.className('slider-handle'));
+    await sliderEl.click();
+    await browser.driver
+      .wait(protractor.ExpectedConditions.visibilityOf(await element(by.css('.tooltip.top.is-open'))), config.waitsFor);
+
+    expect(await element(by.css('.tooltip.top.is-open')).isDisplayed()).toBeTruthy();
+  });
+
+  it('Should not have errors', async () => {
+    await utils.checkForErrors();
+  });
+
+  if (utils.isChrome() && utils.isCI()) {
+    it('Should not visual regress', async () => {
+      const sliderEl = await element(by.className('slider-handle'));
+      await sliderEl.click();
+
+      const mainContent = await element(by.id('maincontent'));
+      await browser.driver.sleep(config.sleep);
+
+      expect(await browser.protractorImageComparison.checkElement(mainContent, 'slider-tooltip-position-top')).toEqual(0);
+    });
+  }
+});